Fix #57: decompose god modules user.rs, admin/main.rs, orm/device.rs
- orm/device.rs (925 → 206): move test module to orm/device/tests.rs
- admin/main.rs (2125 → 204): move 63 tests into per-command modules
  under admin_cli/ where the impl already lived
- api/user.rs (1639): split into api/user/{mod,list,roles}.rs along
  the concerns called out in the issue — OData list/filter/sort/expand
  in list.rs, role-assignment endpoints in roles.rs, rest in mod.rs

No behavior change. All tests pass; lint clean.
