+++ neems-admin/src/admin_cli/site_commands.rs
@@ -373,3 +373,437 @@ pub fn site_edit_impl(
 
     Ok(())
 }
+
+#[cfg(all(test, feature = "test-staging"))]
+#[allow(unused_imports)]
+mod tests {
+    use neems_api::orm::{
+        company::insert_company,
+        site::{get_all_sites, get_site_by_id, get_sites_by_company, insert_site},
+        testing::setup_test_db,
+    };
+
+    use super::*;
+
+    #[test]
+    fn test_handle_site_command_with_conn_ls() {
+        let mut conn = setup_test_db();
+
+        let action = SiteAction::Ls {
+            search_term: None,
+            fixed_string: false,
+            company_id: None,
+        };
+        let result = handle_site_command_with_conn(&mut conn, action, 1);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_handle_site_command_with_conn_add() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        let action = SiteAction::Add {
+            name: "CLI Test Site".to_string(),
+            address: "CLI Test Address".to_string(),
+            latitude: 40.7128,
+            longitude: -74.0060,
+            company_id: company.id.to_string(),
+        };
+        let result = handle_site_command_with_conn(&mut conn, action, 1);
+        assert!(result.is_ok());
+
+        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
+        let found = sites.iter().any(|s| s.name == "CLI Test Site");
+        assert!(found);
+    }
+
+    #[test]
+    fn test_handle_site_command_with_conn_rm() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        insert_site(
+            &mut conn,
+            "Remove This Site".to_string(),
+            "Address".to_string(),
+            40.0,
+            -74.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site");
+
+        let action = SiteAction::Rm {
+            search_term: "Remove This".to_string(),
+            fixed_string: true,
+            yes: true,
+            company_id: None,
+        };
+        let result = handle_site_command_with_conn(&mut conn, action, 1);
+        assert!(result.is_ok());
+
+        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
+        let found = sites.iter().any(|s| s.name == "Remove This Site");
+        assert!(!found);
+    }
+
+    #[test]
+    fn test_site_ls_impl_all() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        insert_site(
+            &mut conn,
+            "Site 1".to_string(),
+            "Address 1".to_string(),
+            40.0,
+            -74.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 1");
+        insert_site(
+            &mut conn,
+            "Site 2".to_string(),
+            "Address 2".to_string(),
+            41.0,
+            -75.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 2");
+
+        let result = site_ls_impl(&mut conn, None, false, None);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_site_ls_impl_with_search() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        insert_site(
+            &mut conn,
+            "Main Office".to_string(),
+            "123 Main St".to_string(),
+            40.0,
+            -74.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 1");
+        insert_site(
+            &mut conn,
+            "Branch Office".to_string(),
+            "456 Branch Ave".to_string(),
+            41.0,
+            -75.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 2");
+
+        let result = site_ls_impl(&mut conn, Some("Main".to_string()), true, None);
+        assert!(result.is_ok());
+
+        let result = site_ls_impl(&mut conn, Some("^Branch".to_string()), false, None);
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_site_ls_impl_with_company_filter() {
+        let mut conn = setup_test_db();
+
+        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
+            .expect("Failed to create company 1");
+        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
+            .expect("Failed to create company 2");
+
+        insert_site(
+            &mut conn,
+            "Site A".to_string(),
+            "Address A".to_string(),
+            40.0,
+            -74.0,
+            company1.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site A");
+        insert_site(
+            &mut conn,
+            "Site B".to_string(),
+            "Address B".to_string(),
+            41.0,
+            -75.0,
+            company2.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site B");
+
+        let result = site_ls_impl(&mut conn, None, false, Some(company1.id));
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_site_add_impl() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        let result = site_add_impl(
+            &mut conn,
+            "New Site".to_string(),
+            "123 New St".to_string(),
+            40.7128,
+            -74.0060,
+            company.id,
+            1,
+        );
+        assert!(result.is_ok());
+
+        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
+        let found = sites.iter().any(|s| s.name == "New Site");
+        assert!(found);
+    }
+
+    #[test]
+    fn test_site_add_impl_duplicate_name_same_company() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        // Create first site
+        let result = site_add_impl(
+            &mut conn,
+            "Duplicate Site".to_string(),
+            "123 Original St".to_string(),
+            40.7128,
+            -74.0060,
+            company.id,
+            1,
+        );
+        assert!(result.is_ok());
+
+        // Try to create second site with same name in same company - should succeed
+        // gracefully
+        let result = site_add_impl(
+            &mut conn,
+            "Duplicate Site".to_string(),
+            "456 Different St".to_string(),
+            41.0,
+            -75.0,
+            company.id,
+            1,
+        );
+        assert!(result.is_ok()); // Now handles duplicates gracefully
+
+        // Verify there's still only one site with this name for this company
+        let sites = get_sites_by_company(&mut conn, company.id).expect("Failed to get sites");
+        let count = sites.iter().filter(|s| s.name == "Duplicate Site").count();
+        assert_eq!(count, 1);
+    }
+
+    #[test]
+    fn test_site_rm_impl() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        insert_site(
+            &mut conn,
+            "Delete Me Site".to_string(),
+            "Address".to_string(),
+            40.0,
+            -74.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site");
+        insert_site(
+            &mut conn,
+            "Keep Me Site".to_string(),
+            "Address".to_string(),
+            41.0,
+            -75.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site");
+
+        let result = site_rm_impl(&mut conn, "Delete Me".to_string(), true, true, None, 1);
+        assert!(result.is_ok());
+
+        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
+        let found_deleted = sites.iter().any(|s| s.name == "Delete Me Site");
+        assert!(!found_deleted);
+
+        let found_kept = sites.iter().any(|s| s.name == "Keep Me Site");
+        assert!(found_kept);
+    }
+
+    #[test]
+    fn test_site_rm_impl_with_company_filter() {
+        let mut conn = setup_test_db();
+
+        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
+            .expect("Failed to create company 1");
+        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
+            .expect("Failed to create company 2");
+
+        insert_site(
+            &mut conn,
+            "Test Site".to_string(),
+            "Address 1".to_string(),
+            40.0,
+            -74.0,
+            company1.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 1");
+        insert_site(
+            &mut conn,
+            "Test Site".to_string(),
+            "Address 2".to_string(),
+            41.0,
+            -75.0,
+            company2.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site 2");
+
+        // Delete only from company1
+        let result =
+            site_rm_impl(&mut conn, "Test Site".to_string(), true, true, Some(company1.id), 1);
+        assert!(result.is_ok());
+
+        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
+        assert_eq!(sites.len(), 1);
+        assert_eq!(sites[0].company_id, company2.id);
+    }
+
+    #[test]
+    fn test_site_edit_impl() {
+        let mut conn = setup_test_db();
+
+        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
+            .expect("Failed to create company 1");
+        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
+            .expect("Failed to create company 2");
+
+        let site = insert_site(
+            &mut conn,
+            "Original Site".to_string(),
+            "Original Address".to_string(),
+            40.0,
+            -74.0,
+            company1.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site");
+
+        // Edit name and address
+        let result = site_edit_impl(
+            &mut conn,
+            site.id,
+            Some("Updated Site".to_string()),
+            Some("Updated Address".to_string()),
+            None,
+            None,
+            None,
+            1,
+        );
+        assert!(result.is_ok());
+
+        let updated_site = get_site_by_id(&mut conn, site.id)
+            .expect("Failed to get updated site")
+            .expect("Site should exist");
+        assert_eq!(updated_site.name, "Updated Site");
+        assert_eq!(updated_site.address, "Updated Address");
+        assert_eq!(updated_site.latitude, 40.0);
+        assert_eq!(updated_site.longitude, -74.0);
+
+        // Edit coordinates and company
+        let result = site_edit_impl(
+            &mut conn,
+            site.id,
+            None,
+            None,
+            Some(41.0),
+            Some(-75.0),
+            Some(company2.id),
+            1,
+        );
+        assert!(result.is_ok());
+
+        let updated_site = get_site_by_id(&mut conn, site.id)
+            .expect("Failed to get updated site")
+            .expect("Site should exist");
+        assert_eq!(updated_site.latitude, 41.0);
+        assert_eq!(updated_site.longitude, -75.0);
+        assert_eq!(updated_site.company_id, company2.id);
+    }
+
+    #[test]
+    fn test_site_edit_impl_nonexistent_site() {
+        let mut conn = setup_test_db();
+
+        let result = site_edit_impl(
+            &mut conn,
+            99999,
+            Some("New Name".to_string()),
+            None,
+            None,
+            None,
+            None,
+            1,
+        );
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_site_edit_impl_nonexistent_company() {
+        let mut conn = setup_test_db();
+
+        let company = insert_company(&mut conn, "Test Company".to_string(), None)
+            .expect("Failed to create company");
+
+        let site = insert_site(
+            &mut conn,
+            "Test Site".to_string(),
+            "Address".to_string(),
+            40.0,
+            -74.0,
+            company.id,
+            120,
+            Some(1),
+        )
+        .expect("Failed to create site");
+
+        let result = site_edit_impl(&mut conn, site.id, None, None, None, None, Some(99999), 1);
+        assert!(result.is_err());
+    }
+}
