+++ neems-admin/src/main.rs
@@ -202,1924 +202,3 @@ async fn get_api_status() -> Result<ApiStatus, Box<dyn std::error::Error>> {
     let api_status: ApiStatus = response.json().await?;
     Ok(api_status)
 }
-
-#[cfg(all(test, feature = "test-staging"))]
-mod tests {
-    use admin_cli::{
-        company_commands::{company_add_impl, company_edit_impl, company_ls_impl, company_rm_impl},
-        device_commands::{device_add_impl, device_edit_impl, device_ls_impl, device_rm_impl},
-        role_commands::{role_add_impl, role_ls_impl},
-        site_commands::{site_add_impl, site_edit_impl, site_ls_impl, site_rm_impl},
-        user_commands::{
-            add_user_impl, change_password_impl, hash_password, list_users_impl, remove_users_impl,
-            user_add_role_impl, user_edit_impl, user_rm_role_impl, user_set_roles_impl,
-        },
-    };
-    use argon2::{Argon2, PasswordHash, PasswordVerifier};
-    #[cfg(feature = "test-staging")]
-    use neems_api::orm::testing::setup_test_db;
-    use neems_api::{
-        models::{CompanyInput, DeviceInput},
-        orm::{
-            company::{get_all_companies, get_company_by_id, get_company_by_name, insert_company},
-            device::{get_all_devices, get_device_by_id, get_devices_by_site, insert_device},
-            role::get_all_roles,
-            site::{get_all_sites, get_site_by_id, get_sites_by_company, insert_site},
-            user::{get_user, get_user_by_email, list_all_users},
-        },
-    };
-
-    use super::*;
-
-    #[test]
-    fn test_handle_user_command_with_conn_add() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        let action = UserAction::Add {
-            email: "cli_test@example.com".to_string(),
-            password: Some("cli_password".to_string()),
-            company_id: company.id.to_string(),
-            totp_secret: Some("cli_totp".to_string()),
-        };
-
-        let result = handle_user_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        // Verify user was created
-        let user = get_user_by_email(&mut conn, "cli_test@example.com")
-            .expect("Failed to get CLI created user")
-            .expect("User should exist");
-        assert_eq!(user.email, "cli_test@example.com");
-        assert_eq!(user.company_id, company.id);
-    }
-
-    #[test]
-    fn test_handle_user_command_with_conn_change_password() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        // Create user first
-        let create_action = UserAction::Add {
-            email: "change_test@example.com".to_string(),
-            password: Some("original".to_string()),
-            company_id: company.id.to_string(),
-            totp_secret: None,
-        };
-        handle_user_command_with_conn(&mut conn, create_action, 1).expect("Failed to create user");
-
-        let original_user = get_user_by_email(&mut conn, "change_test@example.com")
-            .expect("Failed to get user")
-            .expect("User should exist");
-        let original_hash = original_user.password_hash.clone();
-
-        // Change password
-        let change_action = UserAction::ChangePassword {
-            email: "change_test@example.com".to_string(),
-            password: Some("new_password".to_string()),
-        };
-
-        let result = handle_user_command_with_conn(&mut conn, change_action, 1);
-        assert!(result.is_ok());
-
-        // Verify password changed
-        let updated_user = get_user_by_email(&mut conn, "change_test@example.com")
-            .expect("Failed to get updated user")
-            .expect("User should exist");
-        assert_ne!(updated_user.password_hash, original_hash);
-    }
-
-    #[test]
-    fn test_handle_user_command_with_conn_list() {
-        let mut conn = setup_test_db();
-
-        let action = UserAction::Ls { search_term: None, fixed_string: false };
-        let result = handle_user_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_handle_user_command_with_conn_rm() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        let create_action = UserAction::Add {
-            email: "delete_me@example.com".to_string(),
-            password: Some("password1".to_string()),
-            company_id: company.id.to_string(),
-            totp_secret: None,
-        };
-        handle_user_command_with_conn(&mut conn, create_action, 1).expect("Failed to create user");
-
-        let create_action2 = UserAction::Add {
-            email: "keep_me@test.com".to_string(),
-            password: Some("password2".to_string()),
-            company_id: company.id.to_string(),
-            totp_secret: None,
-        };
-        handle_user_command_with_conn(&mut conn, create_action2, 1).expect("Failed to create user");
-
-        let action = UserAction::Rm {
-            search_term: "@example.com".to_string(),
-            fixed_string: true,
-            yes: true,
-        };
-        let result = handle_user_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let remaining_users = list_all_users(&mut conn).expect("Failed to list users");
-        assert_eq!(remaining_users.len(), 1);
-        assert_eq!(remaining_users[0].email, "keep_me@test.com");
-    }
-
-    #[test]
-    fn test_handle_company_command_with_conn_ls() {
-        let mut conn = setup_test_db();
-
-        let action = CompanyAction::Ls { search_term: None, fixed_string: false };
-        let result = handle_company_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_handle_company_command_with_conn_add() {
-        let mut conn = setup_test_db();
-
-        let action = CompanyAction::Add { name: "CLI Test Company".to_string() };
-        let result = handle_company_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let companies = get_all_companies(&mut conn).expect("Failed to get companies");
-        let found = companies.iter().any(|c| c.name == "CLI Test Company");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_handle_company_command_with_conn_rm() {
-        let mut conn = setup_test_db();
-
-        insert_company(&mut conn, "Remove This Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let action = CompanyAction::Rm {
-            search_term: "Remove This".to_string(),
-            fixed_string: true,
-            yes: true,
-        };
-        let result = handle_company_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let companies = get_all_companies(&mut conn).expect("Failed to get companies");
-        let found = companies.iter().any(|c| c.name == "Remove This Company");
-        assert!(!found);
-    }
-
-    #[test]
-    fn test_handle_site_command_with_conn_ls() {
-        let mut conn = setup_test_db();
-
-        let action = SiteAction::Ls {
-            search_term: None,
-            fixed_string: false,
-            company_id: None,
-        };
-        let result = handle_site_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_handle_site_command_with_conn_add() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let action = SiteAction::Add {
-            name: "CLI Test Site".to_string(),
-            address: "CLI Test Address".to_string(),
-            latitude: 40.7128,
-            longitude: -74.0060,
-            company_id: company.id.to_string(),
-        };
-        let result = handle_site_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
-        let found = sites.iter().any(|s| s.name == "CLI Test Site");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_handle_site_command_with_conn_rm() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        insert_site(
-            &mut conn,
-            "Remove This Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site");
-
-        let action = SiteAction::Rm {
-            search_term: "Remove This".to_string(),
-            fixed_string: true,
-            yes: true,
-            company_id: None,
-        };
-        let result = handle_site_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
-        let found = sites.iter().any(|s| s.name == "Remove This Site");
-        assert!(!found);
-    }
-
-    #[test]
-    fn test_hash_password() {
-        let password = "test_password_123";
-        let hash = hash_password(password).expect("Failed to hash password");
-
-        // Verify the hash is valid argon2 format
-        assert!(hash.starts_with("$argon2"));
-
-        // Verify we can verify the password with the hash
-        let argon2 = Argon2::default();
-        let parsed_hash = PasswordHash::new(&hash).expect("Failed to parse hash");
-        assert!(argon2.verify_password(password.as_bytes(), &parsed_hash).is_ok());
-    }
-
-    #[test]
-    fn test_add_user_impl() {
-        let mut conn = setup_test_db();
-
-        // Create a test company first
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        // Test creating a user
-        let result = add_user_impl(
-            &mut conn,
-            "test@example.com",
-            Some("password123".to_string()),
-            company.id,
-            Some("totp_secret".to_string()),
-            1,
-        );
-
-        assert!(result.is_ok());
-
-        // Verify user was created by fetching it
-        let created_user = get_user_by_email(&mut conn, "test@example.com")
-            .expect("Failed to get created user")
-            .expect("User should exist");
-
-        assert_eq!(created_user.email, "test@example.com");
-        assert_eq!(created_user.company_id, company.id);
-        assert_eq!(created_user.totp_secret, Some("totp_secret".to_string()));
-
-        // Verify password was hashed (not stored as plaintext)
-        assert_ne!(created_user.password_hash, "password123");
-        assert!(created_user.password_hash.starts_with("$argon2"));
-    }
-
-    #[test]
-    fn test_add_user_impl_duplicate_email() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        // Create first user
-        add_user_impl(
-            &mut conn,
-            "test@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create first user");
-
-        // Try to create second user with same email - should now succeed gracefully
-        let result = add_user_impl(
-            &mut conn,
-            "test@example.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        );
-
-        assert!(result.is_ok()); // Now handles duplicates gracefully
-    }
-
-    #[test]
-    fn test_change_password_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        // Create a user first
-        add_user_impl(
-            &mut conn,
-            "test@example.com",
-            Some("original_password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let original_user = get_user_by_email(&mut conn, "test@example.com")
-            .expect("Failed to get user")
-            .expect("User should exist");
-        let original_hash = original_user.password_hash.clone();
-
-        // Change password
-        let result = change_password_impl(
-            &mut conn,
-            "test@example.com",
-            Some("new_password".to_string()),
-            1,
-        );
-        assert!(result.is_ok());
-
-        // Verify password was changed
-        let updated_user = get_user_by_email(&mut conn, "test@example.com")
-            .expect("Failed to get updated user")
-            .expect("User should exist");
-
-        assert_ne!(updated_user.password_hash, original_hash);
-        assert!(updated_user.password_hash.starts_with("$argon2"));
-
-        // Verify new password works
-        let argon2 = Argon2::default();
-        let parsed_hash =
-            PasswordHash::new(&updated_user.password_hash).expect("Failed to parse new hash");
-        assert!(argon2.verify_password("new_password".as_bytes(), &parsed_hash).is_ok());
-    }
-
-    #[test]
-    fn test_change_password_impl_nonexistent_user() {
-        let mut conn = setup_test_db();
-
-        let result = change_password_impl(
-            &mut conn,
-            "nonexistent@example.com",
-            Some("password".to_string()),
-            1,
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_list_users_impl_empty() {
-        let mut conn = setup_test_db();
-
-        // Should not panic with empty database
-        let result = list_users_impl(&mut conn, None, false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_list_users_impl_with_users() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        // Create a few users
-        add_user_impl(
-            &mut conn,
-            "user1@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user1");
-        add_user_impl(
-            &mut conn,
-            "user2@example.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user2");
-
-        let result = list_users_impl(&mut conn, None, false);
-        assert!(result.is_ok());
-
-        // Verify users exist
-        let users = list_all_users(&mut conn).expect("Failed to list users");
-        assert_eq!(users.len(), 2);
-        assert_eq!(users[0].email, "user1@example.com");
-        assert_eq!(users[1].email, "user2@example.com");
-    }
-
-    #[test]
-    fn test_list_users_impl_with_regex_search() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "alice@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user1");
-        add_user_impl(
-            &mut conn,
-            "bob@test.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user2");
-        add_user_impl(
-            &mut conn,
-            "charlie@example.org",
-            Some("password3".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user3");
-
-        let result = list_users_impl(&mut conn, Some("example\\.com$".to_string()), false);
-        assert!(result.is_ok());
-
-        let result = list_users_impl(&mut conn, Some("@test".to_string()), false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_list_users_impl_with_fixed_string_search() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "user.with.dots@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user1");
-        add_user_impl(
-            &mut conn,
-            "normaluser@test.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user2");
-
-        let result = list_users_impl(&mut conn, Some(".with.".to_string()), true);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_list_users_impl_invalid_regex() {
-        let mut conn = setup_test_db();
-
-        let result = list_users_impl(&mut conn, Some("[invalid".to_string()), false);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_list_users_impl_no_matches() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "user@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let result = list_users_impl(&mut conn, Some("nonexistent".to_string()), false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_remove_users_impl_with_regex() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "alice@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user1");
-        add_user_impl(
-            &mut conn,
-            "bob@test.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user2");
-        add_user_impl(
-            &mut conn,
-            "charlie@example.org",
-            Some("password3".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user3");
-
-        let result = remove_users_impl(&mut conn, "example\\.com$".to_string(), false, true, 1);
-        assert!(result.is_ok());
-
-        let remaining_users = list_all_users(&mut conn).expect("Failed to list users");
-        assert_eq!(remaining_users.len(), 2);
-        assert_eq!(remaining_users[0].email, "bob@test.com");
-        assert_eq!(remaining_users[1].email, "charlie@example.org");
-    }
-
-    #[test]
-    fn test_remove_users_impl_with_fixed_string() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "user.with.dots@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user1");
-        add_user_impl(
-            &mut conn,
-            "normaluser@test.com",
-            Some("password2".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user2");
-
-        let result = remove_users_impl(&mut conn, ".with.".to_string(), true, true, 1);
-        assert!(result.is_ok());
-
-        let remaining_users = list_all_users(&mut conn).expect("Failed to list users");
-        assert_eq!(remaining_users.len(), 1);
-        assert_eq!(remaining_users[0].email, "normaluser@test.com");
-    }
-
-    #[test]
-    fn test_remove_users_impl_no_matches() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "user@example.com",
-            Some("password1".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let result = remove_users_impl(&mut conn, "nonexistent".to_string(), false, true, 1);
-        assert!(result.is_ok());
-
-        let remaining_users = list_all_users(&mut conn).expect("Failed to list users");
-        assert_eq!(remaining_users.len(), 1);
-    }
-
-    #[test]
-    fn test_remove_users_impl_invalid_regex() {
-        let mut conn = setup_test_db();
-
-        let result = remove_users_impl(&mut conn, "[invalid".to_string(), false, true, 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_change_password_impl_with_provided_password() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        add_user_impl(
-            &mut conn,
-            "password_test@example.com",
-            Some("original_password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let original_user = get_user_by_email(&mut conn, "password_test@example.com")
-            .expect("Failed to get user")
-            .expect("User should exist");
-        let original_hash = original_user.password_hash.clone();
-
-        let result = change_password_impl(
-            &mut conn,
-            "password_test@example.com",
-            Some("new_password".to_string()),
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_user = get_user_by_email(&mut conn, "password_test@example.com")
-            .expect("Failed to get updated user")
-            .expect("User should exist");
-
-        assert_ne!(updated_user.password_hash, original_hash);
-        assert!(updated_user.password_hash.starts_with("$argon2"));
-    }
-
-    #[test]
-    fn test_add_user_impl_with_provided_password() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        let result = add_user_impl(
-            &mut conn,
-            "create_test@example.com",
-            Some("test_password".to_string()),
-            company.id,
-            None,
-            1,
-        );
-        assert!(result.is_ok());
-
-        let created_user = get_user_by_email(&mut conn, "create_test@example.com")
-            .expect("Failed to get created user")
-            .expect("User should exist");
-
-        assert_eq!(created_user.email, "create_test@example.com");
-        assert_eq!(created_user.company_id, company.id);
-        assert!(created_user.password_hash.starts_with("$argon2"));
-    }
-
-    #[test]
-    fn test_company_ls_impl_all() {
-        let mut conn = setup_test_db();
-
-        insert_company(&mut conn, "Test Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        insert_company(&mut conn, "Test Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        let result = company_ls_impl(&mut conn, None, false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_company_ls_impl_with_search() {
-        let mut conn = setup_test_db();
-
-        insert_company(&mut conn, "ACME Corp".to_string(), None)
-            .expect("Failed to create company 1");
-        insert_company(&mut conn, "Tech Solutions".to_string(), None)
-            .expect("Failed to create company 2");
-
-        let result = company_ls_impl(&mut conn, Some("ACME".to_string()), true);
-        assert!(result.is_ok());
-
-        let result = company_ls_impl(&mut conn, Some("^Tech".to_string()), false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_company_add_impl() {
-        let mut conn = setup_test_db();
-
-        let result = company_add_impl(&mut conn, "New Test Company".to_string(), 1);
-        assert!(result.is_ok());
-
-        let companies = get_all_companies(&mut conn).expect("Failed to get companies");
-        let found = companies.iter().any(|c| c.name == "New Test Company");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_company_add_impl_duplicate_name() {
-        let mut conn = setup_test_db();
-
-        // Create first company
-        let result = company_add_impl(&mut conn, "Duplicate Test Company".to_string(), 1);
-        assert!(result.is_ok());
-
-        // Try to create second company with same name - should succeed gracefully
-        let result = company_add_impl(&mut conn, "Duplicate Test Company".to_string(), 1);
-        assert!(result.is_ok()); // Now handles duplicates gracefully
-
-        // Verify there's still only one company with this name
-        let companies = get_all_companies(&mut conn).expect("Failed to get companies");
-        let count = companies.iter().filter(|c| c.name == "Duplicate Test Company").count();
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn test_company_rm_impl_with_cascade() {
-        let mut conn = setup_test_db();
-
-        // Create company with users and sites
-        let company = insert_company(&mut conn, "Delete Me Company".to_string(), None)
-            .expect("Failed to create company");
-        let keep_company = insert_company(&mut conn, "Keep Me Company".to_string(), None)
-            .expect("Failed to create company");
-
-        // Create user in company to be deleted
-        add_user_impl(
-            &mut conn,
-            "user@deleteme.com",
-            Some("password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        // Create user in company to keep
-        add_user_impl(
-            &mut conn,
-            "user@keepme.com",
-            Some("password".to_string()),
-            keep_company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        // Delete company
-        let result = company_rm_impl(&mut conn, "Delete Me".to_string(), true, true, 1);
-        assert!(result.is_ok());
-
-        // Verify company was deleted
-        let companies = get_all_companies(&mut conn).expect("Failed to get companies");
-        let found_deleted = companies.iter().any(|c| c.name == "Delete Me Company");
-        assert!(!found_deleted);
-
-        // Verify other company still exists
-        let found_kept = companies.iter().any(|c| c.name == "Keep Me Company");
-        assert!(found_kept);
-
-        // Verify users were deleted with company
-        let all_users = list_all_users(&mut conn).expect("Failed to get users");
-        let deleted_user_exists = all_users.iter().any(|u| u.email == "user@deleteme.com");
-        assert!(!deleted_user_exists);
-
-        let kept_user_exists = all_users.iter().any(|u| u.email == "user@keepme.com");
-        assert!(kept_user_exists);
-    }
-
-    #[test]
-    fn test_site_ls_impl_all() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        insert_site(
-            &mut conn,
-            "Site 1".to_string(),
-            "Address 1".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 1");
-        insert_site(
-            &mut conn,
-            "Site 2".to_string(),
-            "Address 2".to_string(),
-            41.0,
-            -75.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 2");
-
-        let result = site_ls_impl(&mut conn, None, false, None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_site_ls_impl_with_search() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        insert_site(
-            &mut conn,
-            "Main Office".to_string(),
-            "123 Main St".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 1");
-        insert_site(
-            &mut conn,
-            "Branch Office".to_string(),
-            "456 Branch Ave".to_string(),
-            41.0,
-            -75.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 2");
-
-        let result = site_ls_impl(&mut conn, Some("Main".to_string()), true, None);
-        assert!(result.is_ok());
-
-        let result = site_ls_impl(&mut conn, Some("^Branch".to_string()), false, None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_site_ls_impl_with_company_filter() {
-        let mut conn = setup_test_db();
-
-        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        insert_site(
-            &mut conn,
-            "Site A".to_string(),
-            "Address A".to_string(),
-            40.0,
-            -74.0,
-            company1.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site A");
-        insert_site(
-            &mut conn,
-            "Site B".to_string(),
-            "Address B".to_string(),
-            41.0,
-            -75.0,
-            company2.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site B");
-
-        let result = site_ls_impl(&mut conn, None, false, Some(company1.id));
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_site_add_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let result = site_add_impl(
-            &mut conn,
-            "New Site".to_string(),
-            "123 New St".to_string(),
-            40.7128,
-            -74.0060,
-            company.id,
-            1,
-        );
-        assert!(result.is_ok());
-
-        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
-        let found = sites.iter().any(|s| s.name == "New Site");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_site_add_impl_duplicate_name_same_company() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        // Create first site
-        let result = site_add_impl(
-            &mut conn,
-            "Duplicate Site".to_string(),
-            "123 Original St".to_string(),
-            40.7128,
-            -74.0060,
-            company.id,
-            1,
-        );
-        assert!(result.is_ok());
-
-        // Try to create second site with same name in same company - should succeed
-        // gracefully
-        let result = site_add_impl(
-            &mut conn,
-            "Duplicate Site".to_string(),
-            "456 Different St".to_string(),
-            41.0,
-            -75.0,
-            company.id,
-            1,
-        );
-        assert!(result.is_ok()); // Now handles duplicates gracefully
-
-        // Verify there's still only one site with this name for this company
-        let sites = get_sites_by_company(&mut conn, company.id).expect("Failed to get sites");
-        let count = sites.iter().filter(|s| s.name == "Duplicate Site").count();
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn test_site_rm_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        insert_site(
-            &mut conn,
-            "Delete Me Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site");
-        insert_site(
-            &mut conn,
-            "Keep Me Site".to_string(),
-            "Address".to_string(),
-            41.0,
-            -75.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site");
-
-        let result = site_rm_impl(&mut conn, "Delete Me".to_string(), true, true, None, 1);
-        assert!(result.is_ok());
-
-        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
-        let found_deleted = sites.iter().any(|s| s.name == "Delete Me Site");
-        assert!(!found_deleted);
-
-        let found_kept = sites.iter().any(|s| s.name == "Keep Me Site");
-        assert!(found_kept);
-    }
-
-    #[test]
-    fn test_site_rm_impl_with_company_filter() {
-        let mut conn = setup_test_db();
-
-        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address 1".to_string(),
-            40.0,
-            -74.0,
-            company1.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 1");
-        insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address 2".to_string(),
-            41.0,
-            -75.0,
-            company2.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site 2");
-
-        // Delete only from company1
-        let result =
-            site_rm_impl(&mut conn, "Test Site".to_string(), true, true, Some(company1.id), 1);
-        assert!(result.is_ok());
-
-        let sites = get_all_sites(&mut conn).expect("Failed to get sites");
-        assert_eq!(sites.len(), 1);
-        assert_eq!(sites[0].company_id, company2.id);
-    }
-
-    #[test]
-    fn test_user_edit_impl() {
-        let mut conn = setup_test_db();
-
-        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        // Create a user
-        add_user_impl(
-            &mut conn,
-            "original@example.com",
-            Some("password".to_string()),
-            company1.id,
-            Some("original_totp".to_string()),
-            1,
-        )
-        .expect("Failed to create user");
-
-        let user = get_user_by_email(&mut conn, "original@example.com")
-            .expect("Failed to get user")
-            .expect("User should exist");
-
-        // Edit email
-        let result = user_edit_impl(
-            &mut conn,
-            user.id,
-            Some("updated@example.com".to_string()),
-            None,
-            None,
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_user = get_user(&mut conn, user.id)
-            .expect("Failed to get updated user")
-            .expect("User should exist");
-        assert_eq!(updated_user.email, "updated@example.com");
-        assert_eq!(updated_user.company_id, company1.id);
-
-        // Edit company and TOTP
-        let result = user_edit_impl(
-            &mut conn,
-            user.id,
-            None,
-            Some(company2.id),
-            Some("new_totp".to_string()),
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_user = get_user(&mut conn, user.id)
-            .expect("Failed to get updated user")
-            .expect("User should exist");
-        assert_eq!(updated_user.company_id, company2.id);
-        assert_eq!(updated_user.totp_secret, Some("new_totp".to_string()));
-    }
-
-    #[test]
-    fn test_user_edit_impl_nonexistent_user() {
-        let mut conn = setup_test_db();
-
-        let result =
-            user_edit_impl(&mut conn, 99999, Some("new@example.com".to_string()), None, None, 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_user_edit_impl_nonexistent_company() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        add_user_impl(
-            &mut conn,
-            "user@example.com",
-            Some("password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let user = get_user_by_email(&mut conn, "user@example.com")
-            .expect("Failed to get user")
-            .expect("User should exist");
-
-        let result = user_edit_impl(&mut conn, user.id, None, Some(99999), None, 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_company_edit_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Original Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let result = company_edit_impl(&mut conn, company.id, Some("Updated Company".to_string()));
-        assert!(result.is_ok());
-
-        let updated_company = get_company_by_id(&mut conn, company.id)
-            .expect("Failed to get updated company")
-            .expect("Company should exist");
-        assert_eq!(updated_company.name, "Updated Company");
-    }
-
-    #[test]
-    fn test_company_edit_impl_nonexistent_company() {
-        let mut conn = setup_test_db();
-
-        let result = company_edit_impl(&mut conn, 99999, Some("New Name".to_string()));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_site_edit_impl() {
-        let mut conn = setup_test_db();
-
-        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        let site = insert_site(
-            &mut conn,
-            "Original Site".to_string(),
-            "Original Address".to_string(),
-            40.0,
-            -74.0,
-            company1.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site");
-
-        // Edit name and address
-        let result = site_edit_impl(
-            &mut conn,
-            site.id,
-            Some("Updated Site".to_string()),
-            Some("Updated Address".to_string()),
-            None,
-            None,
-            None,
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_site = get_site_by_id(&mut conn, site.id)
-            .expect("Failed to get updated site")
-            .expect("Site should exist");
-        assert_eq!(updated_site.name, "Updated Site");
-        assert_eq!(updated_site.address, "Updated Address");
-        assert_eq!(updated_site.latitude, 40.0);
-        assert_eq!(updated_site.longitude, -74.0);
-
-        // Edit coordinates and company
-        let result = site_edit_impl(
-            &mut conn,
-            site.id,
-            None,
-            None,
-            Some(41.0),
-            Some(-75.0),
-            Some(company2.id),
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_site = get_site_by_id(&mut conn, site.id)
-            .expect("Failed to get updated site")
-            .expect("Site should exist");
-        assert_eq!(updated_site.latitude, 41.0);
-        assert_eq!(updated_site.longitude, -75.0);
-        assert_eq!(updated_site.company_id, company2.id);
-    }
-
-    #[test]
-    fn test_site_edit_impl_nonexistent_site() {
-        let mut conn = setup_test_db();
-
-        let result = site_edit_impl(
-            &mut conn,
-            99999,
-            Some("New Name".to_string()),
-            None,
-            None,
-            None,
-            None,
-            1,
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_site_edit_impl_nonexistent_company() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            Some(1),
-        )
-        .expect("Failed to create site");
-
-        let result = site_edit_impl(&mut conn, site.id, None, None, None, None, Some(99999), 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_role_add_impl() {
-        let mut conn = setup_test_db();
-
-        let result =
-            role_add_impl(&mut conn, "Test Role".to_string(), Some("A test role".to_string()));
-        assert!(result.is_ok());
-
-        // Verify role was created
-        let roles = get_all_roles(&mut conn).expect("Failed to get roles");
-        let found = roles.iter().any(|r| r.name == "Test Role");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_role_ls_impl() {
-        let mut conn = setup_test_db();
-
-        let result = role_ls_impl(&mut conn, None, false);
-        assert!(result.is_ok());
-
-        // Test with search
-        let result = role_ls_impl(&mut conn, Some("admin".to_string()), false);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_user_add_role_impl() {
-        let mut conn = setup_test_db();
-
-        let company =
-            get_company_by_name(&mut conn, &CompanyInput { name: "Newtown Energy".to_string() })
-                .expect("Failed to query company")
-                .expect("Newtown Energy company should exist");
-
-        add_user_impl(
-            &mut conn,
-            "test@example.com",
-            Some("password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let result = user_add_role_impl(&mut conn, "test@example.com", "newtown-staff");
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_user_set_roles_impl() {
-        let mut conn = setup_test_db();
-
-        let company =
-            get_company_by_name(&mut conn, &CompanyInput { name: "Newtown Energy".to_string() })
-                .expect("Failed to query company")
-                .expect("Newtown Energy company should exist");
-
-        add_user_impl(
-            &mut conn,
-            "test2@example.com",
-            Some("password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        let result =
-            user_set_roles_impl(&mut conn, "test2@example.com", "newtown-admin,newtown-staff");
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_user_rm_role_impl_last_role_fails() {
-        let mut conn = setup_test_db();
-
-        let company =
-            get_company_by_name(&mut conn, &CompanyInput { name: "Newtown Energy".to_string() })
-                .expect("Failed to query company")
-                .expect("Newtown Energy company should exist");
-
-        add_user_impl(
-            &mut conn,
-            "test3@example.com",
-            Some("password".to_string()),
-            company.id,
-            None,
-            1,
-        )
-        .expect("Failed to create user");
-
-        // First assign a role
-        user_add_role_impl(&mut conn, "test3@example.com", "newtown-admin")
-            .expect("Failed to add role");
-
-        // Try to remove the only role - should fail
-        let result = user_rm_role_impl(&mut conn, "test3@example.com", "newtown-admin");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_handle_device_command_with_conn_ls() {
-        let mut conn = setup_test_db();
-
-        let action = DeviceAction::Ls {
-            search_term: None,
-            fixed_string: false,
-            company_id: None,
-            site_id: None,
-        };
-        let result = handle_device_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_handle_device_command_with_conn_add() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create test company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "123 Test St".to_string(),
-            40.7128,
-            -74.0060,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let action = DeviceAction::Add {
-            name: Some("Test Device".to_string()),
-            description: Some("Test Description".to_string()),
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: Some("SN123".to_string()),
-            ip_address: Some("192.168.1.1".to_string()),
-            install_date: Some("2024-01-15 10:30:00".to_string()),
-            company_id: company.id.to_string(),
-            site_id: site.id,
-        };
-        let result = handle_device_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        // Verify device was created
-        let devices = get_all_devices(&mut conn).expect("Failed to get devices");
-        let found = devices.iter().any(|d| d.name == "Test Device");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_handle_device_command_with_conn_rm() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let device_input = DeviceInput {
-            name: Some("Remove This Device".to_string()),
-            description: None,
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input, Some(1)).expect("Failed to create device");
-
-        let action = DeviceAction::Rm {
-            search_term: "Remove This".to_string(),
-            fixed_string: true,
-            yes: true,
-            company_id: None,
-            site_id: None,
-        };
-        let result = handle_device_command_with_conn(&mut conn, action, 1);
-        assert!(result.is_ok());
-
-        let devices = get_all_devices(&mut conn).expect("Failed to get devices");
-        let found = devices.iter().any(|d| d.name == "Remove This Device");
-        assert!(!found);
-    }
-
-    #[test]
-    fn test_device_ls_impl_all() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let device_input1 = DeviceInput {
-            name: Some("Device 1".to_string()),
-            description: None,
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input1, None).expect("Failed to create device 1");
-
-        let device_input2 = DeviceInput {
-            name: Some("Device 2".to_string()),
-            description: None,
-            type_: "Meter".to_string(),
-            model: "Model B".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input2, None).expect("Failed to create device 2");
-
-        let result = device_ls_impl(&mut conn, None, false, None, None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_device_ls_impl_with_search() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let device_input1 = DeviceInput {
-            name: Some("Solar Inverter".to_string()),
-            description: None,
-            type_: "Inverter".to_string(),
-            model: "SUN2000".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input1, None).expect("Failed to create device 1");
-
-        let device_input2 = DeviceInput {
-            name: Some("Battery System".to_string()),
-            description: None,
-            type_: "Battery".to_string(),
-            model: "PowerWall".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input2, None).expect("Failed to create device 2");
-
-        let result = device_ls_impl(&mut conn, Some("Solar".to_string()), true, None, None);
-        assert!(result.is_ok());
-
-        let result = device_ls_impl(&mut conn, Some("^Battery".to_string()), false, None, None);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_device_ls_impl_with_site_filter() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site1 = insert_site(
-            &mut conn,
-            "Site 1".to_string(),
-            "Address 1".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site 1");
-
-        let site2 = insert_site(
-            &mut conn,
-            "Site 2".to_string(),
-            "Address 2".to_string(),
-            41.0,
-            -75.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site 2");
-
-        let device_input1 = DeviceInput {
-            name: Some("Device A".to_string()),
-            description: None,
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site1.id,
-        };
-        insert_device(&mut conn, device_input1, None).expect("Failed to create device A");
-
-        let device_input2 = DeviceInput {
-            name: Some("Device B".to_string()),
-            description: None,
-            type_: "Meter".to_string(),
-            model: "Model B".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site2.id,
-        };
-        insert_device(&mut conn, device_input2, None).expect("Failed to create device B");
-
-        let result = device_ls_impl(&mut conn, None, false, None, Some(site1.id));
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_device_add_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "123 Test St".to_string(),
-            40.7128,
-            -74.0060,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let device_input = DeviceInput {
-            name: Some("New Device".to_string()),
-            description: Some("Device Description".to_string()),
-            type_: "Controller".to_string(),
-            model: "Model X".to_string(),
-            serial: Some("SN456".to_string()),
-            ip_address: Some("192.168.1.100".to_string()),
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-
-        let result = device_add_impl(&mut conn, device_input, 1);
-        assert!(result.is_ok());
-
-        let devices = get_all_devices(&mut conn).expect("Failed to get devices");
-        let found = devices.iter().any(|d| d.name == "New Device");
-        assert!(found);
-    }
-
-    #[test]
-    fn test_device_add_impl_duplicate_name_same_site() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "123 Test St".to_string(),
-            40.7128,
-            -74.0060,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        // Create first device
-        let device_input1 = DeviceInput {
-            name: Some("Duplicate Device".to_string()),
-            description: None,
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        let result = device_add_impl(&mut conn, device_input1, 1);
-        assert!(result.is_ok());
-
-        // Try to create second device with same name in same site - should succeed
-        // gracefully
-        let device_input2 = DeviceInput {
-            name: Some("Duplicate Device".to_string()),
-            description: None,
-            type_: "Meter".to_string(),
-            model: "Model B".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        let result = device_add_impl(&mut conn, device_input2, 1);
-        assert!(result.is_ok()); // Should handle duplicates gracefully
-
-        // Verify there's still only one device with this name for this site
-        let devices = get_devices_by_site(&mut conn, site.id).expect("Failed to get devices");
-        let count = devices.iter().filter(|d| d.name == "Duplicate Device").count();
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn test_device_rm_impl() {
-        let mut conn = setup_test_db();
-
-        let company = insert_company(&mut conn, "Test Company".to_string(), None)
-            .expect("Failed to create company");
-
-        let site = insert_site(
-            &mut conn,
-            "Test Site".to_string(),
-            "Address".to_string(),
-            40.0,
-            -74.0,
-            company.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site");
-
-        let device_input1 = DeviceInput {
-            name: Some("Delete Me Device".to_string()),
-            description: None,
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input1, Some(1)).expect("Failed to create device 1");
-
-        let device_input2 = DeviceInput {
-            name: Some("Keep Me Device".to_string()),
-            description: None,
-            type_: "Meter".to_string(),
-            model: "Model B".to_string(),
-            serial: None,
-            ip_address: None,
-            install_date: None,
-            company_id: company.id,
-            site_id: site.id,
-        };
-        insert_device(&mut conn, device_input2, Some(1)).expect("Failed to create device 2");
-
-        let result = device_rm_impl(&mut conn, "Delete Me".to_string(), true, true, None, None, 1);
-        assert!(result.is_ok());
-
-        let devices = get_all_devices(&mut conn).expect("Failed to get devices");
-        let found_deleted = devices.iter().any(|d| d.name == "Delete Me Device");
-        assert!(!found_deleted);
-
-        let found_kept = devices.iter().any(|d| d.name == "Keep Me Device");
-        assert!(found_kept);
-    }
-
-    #[test]
-    fn test_device_edit_impl() {
-        let mut conn = setup_test_db();
-
-        let company1 = insert_company(&mut conn, "Company 1".to_string(), None)
-            .expect("Failed to create company 1");
-        let company2 = insert_company(&mut conn, "Company 2".to_string(), None)
-            .expect("Failed to create company 2");
-
-        let site1 = insert_site(
-            &mut conn,
-            "Site 1".to_string(),
-            "Address 1".to_string(),
-            40.0,
-            -74.0,
-            company1.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site 1");
-
-        let site2 = insert_site(
-            &mut conn,
-            "Site 2".to_string(),
-            "Address 2".to_string(),
-            41.0,
-            -75.0,
-            company2.id,
-            120,
-            None,
-        )
-        .expect("Failed to create site 2");
-
-        let device_input = DeviceInput {
-            name: Some("Original Device".to_string()),
-            description: Some("Original Description".to_string()),
-            type_: "Sensor".to_string(),
-            model: "Model A".to_string(),
-            serial: Some("SN123".to_string()),
-            ip_address: Some("192.168.1.1".to_string()),
-            install_date: None,
-            company_id: company1.id,
-            site_id: site1.id,
-        };
-        let device =
-            insert_device(&mut conn, device_input, Some(1)).expect("Failed to create device");
-
-        // Edit name and description
-        let result = device_edit_impl(
-            &mut conn,
-            device.id,
-            Some("Updated Device".to_string()),
-            Some("".to_string()), // Clear description
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_device = get_device_by_id(&mut conn, device.id)
-            .expect("Failed to get updated device")
-            .expect("Device should exist");
-        assert_eq!(updated_device.name, "Updated Device");
-        assert_eq!(updated_device.description, None);
-        assert_eq!(updated_device.type_, "Sensor");
-        assert_eq!(updated_device.model, "Model A");
-
-        // Edit company and site
-        let result = device_edit_impl(
-            &mut conn,
-            device.id,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            Some(company2.id),
-            Some(site2.id),
-            1,
-        );
-        assert!(result.is_ok());
-
-        let updated_device = get_device_by_id(&mut conn, device.id)
-            .expect("Failed to get updated device")
-            .expect("Device should exist");
-        assert_eq!(updated_device.company_id, company2.id);
-        assert_eq!(updated_device.site_id, site2.id);
-    }
-
-    #[test]
-    fn test_device_edit_impl_nonexistent_device() {
-        let mut conn = setup_test_db();
-
-        let result = device_edit_impl(
-            &mut conn,
-            99999,
-            Some("New Name".to_string()),
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            1,
-        );
-        assert!(result.is_err());
-    }
-}
